deleting registries on uninstall + fixes check_user_permission

+++ src/actions/uninstall/uninstall_folder.sh
@@ -16,21 +16,30 @@
 # Uninstall folders and associated groups on IAC_FOLDERS
 
 uninstall_folder() {
-  for i in $(cat "$IAC_FOLDERS" | sed 1d)
-  do
-  # echo "$i"
-    folderName=$(echo "${i}" | cut -d $SEP -f 1)
-    group=$(echo "${i}"  | cut -d $SEP -f 3)
-    group=$(uppercase "${group}")
- 
-    $( check_system_folder "${IAC_PATH}${folderName}" ) && {
-      echo 
-      echo -n "Are you sure to remove folder ${IAC_PATH}${folderName} and group ${group} ?: "
-      echo 
-      read confirm
+  check_user_permission || {
+    echo "Could not add folder and group: Permission denied. Are you root?"
+    exit 1
+  }
+  
+  echo 
+  echo -n "Are you sure to remove folder ${IAC_PATH}${folderName} and group ${group} ?: "
+  echo 
+  read confirm
 
-      if [ "$confirm" == "y" ] || [ "$confirm" == "Y" ];then
+  if [ "$confirm" == "y" ] || [ "$confirm" == "Y" ];then
+
+    for i in $(cat "$IAC_FOLDERS" | sed 1d)
+    do
+      folderName=$(echo "${i}" | cut -d $SEP -f 1)
+      group=$(echo "${i}"  | cut -d $SEP -f 3)
+      group=$(uppercase "${group}")
+  
+      $( check_system_folder "${IAC_PATH}${folderName}" ) && {
+
         rm -r "${IAC_PATH}${folderName}"
+
+        delete_registry "$folderName" "$IAC_FOLDERS"
+
         echo
         echo "Folder $folderName was removed."
         echo 
@@ -41,17 +50,9 @@ uninstall_folder() {
           echo "Group $group was removed."
           echo 
         }
-      else
-        echo
-        echo "Folder $folderName kept in OS."
-        echo 
-        echo  
-        echo "Group $group kept in OS."
-        echo 
-      fi
-    }
+      }
+    done
 
-  done
-  
+  fi
   exit 0
 }
